ex02/requirements.txt
# ex02 사내 CRUD 시스템 의존성
# Python 3.10+ 필요

# 웹 프레임워크
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# 템플릿 엔진
jinja2>=3.1.0

# 폼 파싱 (multipart/form-data)
python-multipart>=0.0.18

# PostgreSQL 드라이버
psycopg2-binary>=2.9.9

# 환경 변수 관리
python-dotenv>=1.0.0

# 데이터 검증
pydantic>=2.10.0
